Separate aes part from encrypted_file_mapping
encrypted_file_mapping.hpp needs to be inlined and exposed to android
build.
