+++ src/realm/util/encrypted_file_mapping.cpp
@@ -17,6 +17,8 @@
  * from Realm Incorporated.
  *
  **************************************************************************/
+
+#include <realm/util/aes_cryptor.hpp>
 #include <realm/util/encrypted_file_mapping.hpp>
 #include <realm/util/file_mapper.hpp>
 #include <realm/utilities.hpp>
